# docs/swagger.yaml
basePath: /
definitions:
...
  input.MenuButtonsReq:
    properties:
      parentId:
        description: 菜单ID（作为父级ID进行查询）
        example: 0
        type: integer
    required:
    - parentId
    type: object
  input.MenuDeleteReq:
    properties:
      id:
        description: 菜单ID
        example: 1
        type: integer
    required:
    - id
    type: object
  input.MenuSaveReq:
    properties:
      code:
        description: 菜单标识
        example: Base
        minLength: 1
        type: string
      component:
        description: 组件路径
        example: /src/list/list.vue
        type: string
      enable:
        description: 状态
        example: true
        type: boolean
      icon:
        description: 菜单图标
        example: i-fe:list
        type: string
      id:
        description: 菜单ID
        example: 1
        type: integer
      keepAlive:
        description: 保持活跃
        example: false
        type: boolean
      layout:
        description: 布局
        example: full
        type: string
      name:
        description: 菜单名称
        example: 基础菜单
        minLength: 1
        type: string
      order:
        description: 排序（从小到大）
        example: 0
        type: integer
      parentId:
        description: 父级ID
        example: 0
        type: integer
      path:
        description: 菜单路径
        example: /path/url
        type: string
      show:
        description: 显示状态
        example: true
        type: boolean
      type:
        description: 菜单类型
        enum:
        - BUTTON
        - MENU
        example: MENU
        minLength: 1
        type: string
    required:
    - code
    - name
    - type
    type: object
  input.MenuToggleReq:
    properties:
      id:
        description: 菜单ID
        example: 1
        type: integer
    required:
    - id
    type: object
  input.MenuValidateReq:
    properties:
      path:
        description: 菜单路径
        example: /path/url
        minLength: 1
        type: string
    required:
    - path
    type: object
...
  resp.MenuButtonsResp:
    properties:
      pageData:
        description: 当前页码数据
        items:
          $ref: '#/definitions/resp.MenuItem'
        type: array
      total:
        description: 总计条数
        example: 100
        type: integer
    type: object
  resp.MenuItem:
    properties:
      children:
        description: 子菜单
        items:
          $ref: '#/definitions/resp.MenuItem'
        type: array
      code:
        description: 菜单标识
        example: Base
        type: string
      component:
        description: 组件路径
        example: /src/list/list.vue
        type: string
      description:
        description: 描述
        type: string
      enable:
        description: 状态
        example: true
        type: boolean
      icon:
        description: 菜单图标
        example: i-fe:list
        type: string
      id:
        description: 菜单ID
        example: 1
        type: integer
      keepAlive:
        description: 保持活跃
        example: false
        type: boolean
      layout:
        description: 布局
        example: full
        type: string
      method:
        description: 方法
        type: string
      name:
        description: 菜单名称
        example: 基础菜单
        type: string
      order:
        description: 排序（从小到大）
        example: 0
        type: integer
      parentId:
        description: 父级ID
        example: 0
        type: integer
      path:
        description: 菜单路径
        example: /path/url
        type: string
      redirect:
        description: 重定向
        type: string
      show:
        description: 显示状态
        example: true
        type: boolean
      type:
        description: 菜单类型
        enum:
        - BUTTON
        - MENU
        example: MENU
        type: string
    type: object
  resp.MenuListResp:
    properties:
      menu:
        description: 菜单列表
        items:
          $ref: '#/definitions/resp.MenuItem'
        type: array
    type: object
  resp.MenuValidateResp:
    properties:
      has:
        description: 是否存在
        example: true
        type: boolean
    type: object
...
  response.Response:
    properties:
      code:
        description: 响应code
        example: 0
        type: integer
      data:
        description: 响应数据
      msg:
        description: 响应信息
        example: success
        type: string
    type: object
info:
  contact:
    email: junjie.he.925@gmail.com
    name: API 支持
  description: OneAdmin 系统接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: OneAdmin API
  version: "1.0"
paths:
...
  /api/admin/menu/buttons:
    post:
      description: 根据菜单 ID 获取其下的按钮权限列表（按钮类型菜单），用于前端页面按钮级权限控制
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 查询参数（父级菜单 ID）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.MenuButtonsReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.MenuButtonsResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单下的按钮权限
      tags:
      - 菜单
  /api/admin/menu/delete:
    post:
      description: 根据菜单 ID 删除菜单（通常不允许删除存在子菜单或已绑定权限的菜单）
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 删除参数（菜单 ID）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.MenuDeleteReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - 菜单
  /api/admin/menu/list:
    post:
      description: 获取系统完整菜单树结构（包含菜单及按钮权限），用于前端动态路由生成和权限控制
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.MenuListResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单列表
      tags:
      - 菜单
  /api/admin/menu/save:
    post:
      description: 根据是否传入 ID 创建或更新菜单信息：未传 ID 时为创建，传入 ID 时为更新，支持菜单及按钮类型（MENU / BUTTON），用于构建系统菜单结构和权限控制
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 菜单保存参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.MenuSaveReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建或更新菜单
      tags:
      - 菜单
  /api/admin/menu/toggle:
    post:
      description: 根据菜单 ID 切换菜单的启用状态（启用/禁用），常用于后台列表中的快速开关操作
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 切换参数（菜单 ID）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.MenuToggleReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 切换菜单启用状态
      tags:
      - 菜单
  /api/admin/menu/validate:
    post:
      description: 根据菜单路径校验菜单是否已存在，常用于前端需要对菜单进行校验时
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 校验参数（菜单路径）
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.MenuValidateReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.MenuValidateResp'
              type: object
      security:
      - BearerAuth: []
      summary: 校验菜单路径是否存在
      tags:
      - 菜单
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"